feat(web): separate login handler into custom hook
