+++ apps/web/src/features/login/model/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useLogin } from './useLogin';
